Add "Duty Loot Preview" Feature
This adds a Duty Loot Preview window, accessible via a button
on the Duty screen, that shows the loot dropped by a duty.

I wasn't sure if you would be comfortable adding a dependency on
LuminaSupplemental, but I figured the easiest way to check would be
with a PR. Let me know what you think.