Add "Duty Loot Preview" Feature (#60)
* Add "Duty Loot Preview" Feature

This adds a Duty Loot Preview window, accessible via a button
on the Duty screen, that shows the loot dropped by a duty.

I wasn't sure if you would be comfortable adding a dependency on
LuminaSupplemental, but I figured the easiest way to check would be
with a PR. Let me know what you think.

* Style adjustments

* PR Feedback

- Address Style Issues
- Add "No Loot Data Found" message with a third-party data disclaimer
- Scroll to start on list change
- Load data async

Note: For the async loading it was flickering since it's so fast, so I added a small delay to showing the label.

* Final style fixes

---------

Co-authored-by: MidoriKami <mi_kami@outlook.com>